move serde tests
